Added hidden command for creating a key-table
 - takes in a named expression specifying the keys of the new table
 - takes named aggregator expressions specifying how the columns should be computed
